restructure codebase, split into multiple files
