+++ coolpuppy/lib/util.py
